extracts common methods for preprint and publication compliance
