+++ apps/crn-server/scripts/opensearch/publication-compliance.ts
@@ -1,23 +1,9 @@
-import {
-  awsRegion,
-  googleApiCredentialsSecretId,
-  opensearchAnalyticsSpreadsheetUrl,
-} from '../../src/config';
-import {
-  extractSpreadsheetIdFromUrl,
-  getJWTCredentialsFactory,
-  readGoogleSheetsData,
-  readGoogleSheetsDataLocal,
-} from '@asap-hub/server-common';
-import TeamController from '../../src/controllers/team.controller';
-import { getTeamDataProvider } from '../../src/dependencies/team.dependencies';
+import { fetchAllTeams, readComplianceData } from './shared-utils';
 
 import {
   PUBLICATION_COMPLIANCE_ALL_TIME_SHEET_NAME,
   PUBLICATION_COMPLIANCE_LAST_12_MONTHS_SHEET_NAME,
-  SHEET_RANGE,
   PUBLICATION_COMPLIANCE_HEADER_MAPPINGS,
-  BATCH_SIZE,
 } from './constants';
 import { MetricObject } from './types';
 
@@ -26,62 +12,11 @@ interface SpreadsheetRow {
   [key: string]: unknown;
 }
 
-const fetchAllTeams = async () => {
-  try {
-    const teamController = new TeamController(getTeamDataProvider());
-
-    const initialResponse = await teamController.fetch({
-      take: BATCH_SIZE,
-      skip: 0,
-    });
-
-    const allTeams = [...initialResponse.items];
-    const total = initialResponse.total;
-
-    for (let skip = BATCH_SIZE; skip < total; skip += BATCH_SIZE) {
-      const response = await teamController.fetch({
-        take: BATCH_SIZE,
-        skip,
-      });
-      allTeams.push(...response.items);
-    }
-
-    return allTeams;
-  } catch (error) {
-    console.warn(
-      'Failed to fetch teams from Contentful, using empty team list:',
-      error,
-    );
-    return [];
-  }
-};
-
 const readPublicationComplianceData = async (
   sheetName: string,
   environment: 'local' | 'production' = 'production',
 ): Promise<SpreadsheetRow[]> => {
-  const spreadsheetId = extractSpreadsheetIdFromUrl(
-    opensearchAnalyticsSpreadsheetUrl,
-  );
-
-  if (environment === 'local') {
-    const rawData = await readGoogleSheetsDataLocal(
-      spreadsheetId,
-      `${sheetName}!${SHEET_RANGE}`,
-    );
-    return rawData as SpreadsheetRow[];
-  }
-
-  const getJWTCredentials = getJWTCredentialsFactory({
-    googleApiCredentialsSecretId,
-    region: awsRegion,
-  });
-
-  const rawData = await readGoogleSheetsData(getJWTCredentials, {
-    spreadsheetId,
-    range: `${sheetName}!${SHEET_RANGE}`,
-  });
-
+  const rawData = await readComplianceData(sheetName, environment);
   return rawData as SpreadsheetRow[];
 };
